Fix C++11 support detection for Visual Studio compilers
Simply checking on the value of the __cplusplus preprocessor variable
when using MSVC is not reliable to detect C++11 support. Checking on
the definition of _MSVC_LANG one is more robust as it was not the case
before Visual Studio 2015 which is the first release to fully implement
C++11 specifications.